requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
discord.py==0.16.12
multidict==4.5.2
soupsieve==1.9.2
websockets==3.4
